\name{eDNAPlus-package}
\alias{eDNAPlus-package}
\alias{eDNAPlus}
\docType{package}
\title{\packageTitle{eDNAPlus}}
\description{\packageDescription{eDNAPlus}}
\section{Package Content}{\packageIndices{eDNAPlus}}
\author{\packageAuthor{eDNAPlus}}
\section{Maintainer}{\packageMaintainer{eDNAPlus}}
\keyword{package}
