% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotBiomassCoefficients}
\alias{plotBiomassCoefficients}
\title{Biomass covariates coefficients plot}
\usage{
plotBiomassCoefficients(modelResults, cov_num = 1, idxSpecies)
}
\arguments{
\item{modelResults}{Output of the function \code{runEDNA}}

\item{cov_num}{Index of the covariate to plot}

\item{idxSpecies}{Optional indexes of the species to plot. If missing, the significant species are plotted.}
}
\value{
A credible intervals plot with the species on the columns
}
\description{
Plot the covariate coefficients of the biomasses amount
}
\examples{

plotBiomassCoefficients(sampleResults)

}
