% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotBiomassesMap}
\alias{plotBiomassesMap}
\title{Biomasses distribution maps}
\usage{
plotBiomassesMap(modelResults, datapoly, idxSpecies)
}
\arguments{
\item{modelResults}{Output of the function \code{runEDNA}}

\item{datapoly}{Polygon file of the map}

\item{idxSpecies}{Index of the species to plot}
}
\value{
A plot with the map of the biomasses, where a darker color signals higher biomass
}
\description{
Plot the maps of the biomasses
}
\examples{

plotBiomassCoefficients(sampleResults)

}
