% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Species correlation matrix}
\usage{
plotCorrelationMatrix(modelResults, idxSpecies, numSigCorrs = 0)
}
\arguments{
\item{modelResults}{Output of the function \code{runEDNA}}

\item{idxSpecies}{Optional indexes of the species to plot. If missing, the significant correlations are plotted.}

\item{numSigCorrs}{Number of significant correlation to plot}
}
\value{
A plot of the median correlation matrix for the species selected
}
\description{
Plot the species correlation matrix
}
\examples{

plotCorrelationMatrix(sampleResults)

}
