% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotFalsePositiveProbability}
\alias{plotFalsePositiveProbability}
\title{Probability plot of false positive}
\usage{
plotFalsePositiveProbability(modelResults, idxSpecies)
}
\arguments{
\item{modelResults}{Output of the function \code{runEDNA}}
}
\value{
Credible intervals plot of the false positive probabilities
}
\description{
Plot the credible intervals of the false positive probabilities
}
\examples{

plotTruePositiveProbability(sampleResults)

}
