% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotTruePositiveProbability}
\alias{plotTruePositiveProbability}
\title{Probability plot of true positive}
\usage{
plotTruePositiveProbability(modelResults, idxSpecies)
}
\arguments{
\item{modelResults}{Output of the function \code{runEDNA}}
}
\value{
Credible intervals plot of the true positive probabilities
}
\description{
Plot the credible intervals of the true positive probability
}
\examples{

plotTruePositiveProbability(sampleResults)

}
