% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodel.R
\name{runEDNA}
\alias{runEDNA}
\title{Model fitting}
\usage{
runEDNA(
  data,
  priors = NULL,
  jointSpecies = F,
  paramsUpdate = NULL,
  MCMCparams,
  paramsToSave = NULL
)
}
\arguments{
\item{data}{A list containing the data. Detailed informations are given in the vignette. The elements are:
 \itemize{

 \item \code{PCR_table}: A data.frame with the reads of the wildlife species. 
 \item \code{infos}: A data.frame with the samples informations.
 \item \code{X_z}: Covariate matrix for the biomass abundance
 \item \code{X_w}: Covariate matrix for the biomass detection
 \item \code{PCR_spike}: A data.frame with the reads of the spike-in species. 
 }}

\item{priors}{List of prior parameters. If not specified, default values are used.}

\item{jointSpecies}{Does the model include species correlation?}

\item{MCMCparams}{A list with components:
\itemize{
   \item \code{nchain}: Number of chains
   \item \code{nburn}: Number of burn-in iterations
   \item \code{niter}: Number of iterations
   \item \code{nthin}: Number of iterations to thin
}}

\item{paramsToSave}{A list with components the elements of which the MCMC output need to be saved.}
}
\value{
A list with components:

\itemize{

 \item \code{params_output} A list with the MCMC output of each parameter specified in the \code{paramesToSave} parameter
     
 
 }
}
\description{
This function fits the occupancy model of Diana et al. (2021). 
Note that in the following the parameters are described with the notations used in the paper.
}
\examples{

modelResults <- runEDNA(data, 
                        priors,
                        jointSpecies = T,
                        MCMCparams)  

}
